[enhancement] update broadcast join cost algorithm

+++ fe/fe-core/src/test/java/org/apache/doris/planner/DistributedPlannerTest.java
@@ -158,7 +158,7 @@ public class DistributedPlannerTest {
         String plan = planner.getExplainString(fragments, new ExplainOptions(false, false));
         Assert.assertEquals(1, StringUtils.countMatches(plan, "INNER JOIN (BROADCAST)"));
 
-        long originThreshold = ctx.getSessionVariable().autoBroadcastJoinThreshold;
+        double originThreshold = ctx.getSessionVariable().autoBroadcastJoinThreshold;
         try {
             ctx.getSessionVariable().autoBroadcastJoinThreshold = -1L;
             stmtExecutor = new StmtExecutor(ctx, sql);
